added processing, training and testing on external bible datasets

+++ src/data/build_spacy_objects.py
@@ -4,6 +4,7 @@ import pandas as pd
 import csv
 import sys
 import pickle
+from tqdm import tqdm
 
 
 def build(infile_path, nlp):
@@ -31,7 +32,7 @@ def build(infile_path, nlp):
 
         X = []
 
-        for instance in dataset:
+        for instance in tqdm(dataset):
 
             x = nlp(instance['sentence'])
             X.append(x.to_bytes())
@@ -42,6 +43,7 @@ def build(infile_path, nlp):
 if __name__ == '__main__':
     #TODO write this stuff in the readme, makedir if dir is missing.
     language = sys.argv[1]
+    todo_external = sys.argv[2]
     save_dir = 'data/interim/' + language
     data_path = 'data/raw/' + language
     file_paths = os.listdir(data_path)
@@ -63,13 +65,24 @@ if __name__ == '__main__':
         nlp = spacy.load('fr_core_news_md')
     else:
         raise ValueError("Language specified ({}) not supported.".format(language))
+    
+    if todo_external:
+        data_path = "data/raw/external_datasets/"
+        file_paths = os.listdir(data_path)
+        save_dir = "/cwi/data/interim/external_datasets/"
+        os.makedirs(save_dir,exist_ok=True)
+
 
     print('Building spaCy objects for {} language'.format(language))
 
     for file_path in file_paths:
+        interim_save_path = save_dir + "/" + file_path.split(".")[0] + "-spacy-objs.pkl"
+        if not file_path.endswith(".tsv") or os.path.exists(interim_save_path):
+            print(f"Skipping file {file_path}")
+            continue
         print("Building spaCy objects for ", file_path)
 
         X = build(data_path + '/' + file_path, nlp)
 
-        with open(save_dir + "/" + file_path.split(".")[0] + "-spacy-objs.pkl", "wb") as out_file:
+        with open(interim_save_path, "wb") as out_file:
             pickle.dump(X, out_file)
